Add the ability to Listen to Events on pure Script Classes as long as they inherit the ScriptObject interface. This adds an additional optional parameter to to SubscribeToEvent in the following manner:
    SubscribeToEvent("Update","HandleUpdate",@this);

This required a change to how we store the pointer to the ScriptInstance class, since the object user data pointer is lost when moving between script classes, the pointer to the listener for ScriptInstance and ScriptFile has been moved to Context and Execute has been modified so that everytime a function is executed the Context UserData Ptr is updated.

This allowed another change that means a ScriptFile can now handle events that are triggered by a SubscribeToEvents from inside a script object since when calling AddEventReciever there is no need to know and cast to the base class ScriptEventListener can be used as its a pure virtual function.