Separated the Spline Mathematical aspect from the Component into a new Spline class under Math. No angelscript or lua bindings as of yet as I'm unsure how to go about adding the templated class in a suitable manner.
Old Spline class has been renamed to SplinePath as is build using child nodes of its parent node. To facilitate the functionality a method has been added to Node to allows you to specify the index position of where to add the child Node as that impacts the curve.  DebugGeometry was added which shows the nodes that currently make up the SplinePath and the curve.

Added CScriptArray to PODVector convertor.