Add a simple Android samples launcher activity.
The launcher activity will only be shown when there are actually more samples built into the deployed apk. If no other samples then the launcher will just launch Urho3DPlayer without prompting. Otherwise a list view is presented in the launcher activity to let user to choose which sample library to be loaded.
Due to unloading of shared library is not supported from Java side, currently once a sample library is being loaded, the launcher cannot unload the library and let user to choose other samples. Until a workaround is found, user would have to perform a long pressed on a home button, close the app, and restart it again from scratch.