Remove LuaJIT specific definition as it is not needed.
Since we are not installing LuaJIT directly but leave it in Urho3D's Bin directory, it should not be necessary to alter the search path to include LUA_XROOT (in v2.0.2) or LUA_ROOT (in v2.0.3). Similarly because of that, we also do not need to care about LuaJIT new support for MULTILIB.

This change should remove the incorrectly included hardcoded-value in the Urho3D.pc for the Raspberry Pi platform as well.