TmxFile2D: Fix determination of object type
TmxObjectGroup2D::Load was using the presence of the "width" and
"height" attributes for deciding whether an object was a rectangle,
ellipse or something else. This is wrong, and the approach breaks for
maps saved in more recent versions of Tiled. This is because the width
and height are only saved when they differ from the default (0), and
because the width and height are now also used to store the size of tile
objects.

So for maps saved with Tiled >= 0.12, this change:

* Fixes identification of rectangle and ellipse objects with 0 width and
  height (which was leading to the entire map failing to load).
* Fixes identification of tile objects.