Travis CI: API documentation update at 2015-09-11 13:17:42 UTC. [ci package]
Commit: https://github.com/urho3d/Urho3D/commit/7356bdecfbe3ea15822c7f4a3e073eacc6b75776

Message: Add API to Animation / AnimationTrack to create tracks and keyframes and expose it to script. Change AnimationTracks to be stored in a HashMap so that removal and creation is safe without invalidating pointers. Changed the AnimationTriggerPoint API to resemble the AnimationTrack API, and change it to be handled as a value type in AngelScript API. Closes #878.