Hide UIElement3D from public API, clean up UIElement3D and UIComponent.
`UIElement3D` is no longer in public API. This prompted addition of `UIComponent::SetViewportIndex(unsigned index)` API for controlling which viewport is used for UI input mapping. If component does not belong to a node then index will be cached and used when component is added to a node. Not setting viewport index or setting invalid index will result in use of viewport at index 0.

`UIElement3D` no longer holds `UIComponent` pointer. Instead it holds `Node` pointer and obtains `Scene` and `StaticModel` directly from the model. This allows model to be removed and re-added without breaking. Note that after re-adding (likely new) `StaticModel` it is user's responsibility to set material returned by `UIComponent` to a new `StaticModel`.

`UIElement3D` no longer caches model pointer if component does not own the model. As a result of that flag `isStaticModelOwned_` was removed.

Swapped protected members and methods in `UIComponent` to fit style guidelines.