Layer and Position.z in 2D games
Hello everyone, I recently used `StaticSprite2D` to study the relationship between `SetLayer` and` position.z` and found that `z` has priority over `SetLayer.` Layers only work when `z` is equal. I looked at the unity engine and found that it was the opposite. It was to consider layer first and then consider z. Which kind of order is universal? should we be consistent with unity?