Fix compiler warning:
        Resource/JSONValue.h:200:110: warning: reference to stack memory associated with parameter 'defaultValue' returned [-Wreturn-stack-address]
        const String& GetString(String defaultValue = String::EMPTY) const { return IsString() ? *stringValue_ : defaultValue;}

Note: This is also a series problem as the parameter is a stack variable and will be destroyed on
the function return. This returning a reference to this value is returning a reference to a value
that has been destroyed.

Changing the parameter to also be a `const String&` does not change the meaning or usage of the
function.