Fix compiler warning:
Compiler Warning was:

````
    Resource/JSONValue.h:200:110: warning: reference to stack memory associated with parameter 'defaultValue' returned [-Wreturn-stack-address]
    const String& GetString(String defaultValue = String::EMPTY) const { return IsString() ? *stringValue_ : defaultValue;}
````

Now the trouble here is that `String::EMPTY` is a constant value. So we can't just use a reference.

Possible Solutions:

* Pass defaultValue as const reference.
* Pass defaultValue as reference.
* Pass defaultValue by value.
* Pass a type of reference but value keep local copy
* Return by value

As pointed out by `iSLC` is not usable as this allows temporary variables, which may have been destroyed.

The problem here is that `String::EMPTY` has a type `const reference` and thus can not be used as a default value.

This was the original bug. The problem here is that we are returning a reference and if we have a default value then you are returning a reference to a parameter that will be destroyed on function exit.

We could potentially keep a local copy in the function .

1. by using a static function variable. We can return a reference to this variable as it will live an appropriate amount of time. The problem here is that the default value only lives until the next call to `GetString()` at which point it will be over-riden. Not all objects will share the same variable so a call to any `GetString()` will change the value. If a user kept a reference to the value this effectively causes the value to change behind the scenes.

2. by using a member variable. Again it only last until the `GetString()` might be slightly longer as the `GetString()` would have to be called on the same object. But this alters the state of the object which long term is undesirable.

This will allow the parameter to be passed by constant reference. It has the same affect as returning by const reference. The trouble is that it will require a copy to return the value so it becomes less efficient.

The only way to solve the problem is to return by value. This will be less efficient but it will be correct.

````
String GetString(String const& defaultValue = String::EMPTY) const { return IsString() ? *stringValue_ : defaultValue;}
````

There are situations where we can still return by `const reference` if we know both alternatives can still be maintained. So in the situation where no defaultValue is provided by the user we know that String::Empty is always available so we can return that or the original value by const reference.

````
String GetString(String const& defaultValue) const { return IsString() ? *stringValue_ : defaultValue;}
const String& GetString() const { return IsString() ? *stringValue_ : String::EMPTY;}
````

Since GetString() returns a const reference why can we simply not use "Pass defaultValue as reference" as defined above but cast away the `const` for the default value.

````
    const String& GetString(String& defaultValue = const_cast< String & >(String::EMPTY)) const { return IsString() ? *stringValue_ : defaultValue; }
````

The problem here is that we can not use temporary object or pass as default values a const String.

````
    const String  myDefault = "a";
    object.GetString(myDefault);        // Fail
    object.GetString(String("Temp"));   // Fail
````