Update PBRLitSolid.hlsl
Removes a line that does not match the GLSL version of the code, and can make environment mapping to dark when specular color is not full white in the HLSL version.
It may be that it is the GLSL version that needs correcting to make it consistent with the HLSL version, but I'm not really sure why the HLSL version of this code would need to effectively multiply a specular color by itself.  Of course, if material specular colour happens to be pure white then there is no difference between them.